LayoutTests/sputnik/Conformance/12_Statement/12.8_The_break_Statement/S12.8_A1_T3-expected.txt

CONSOLE MESSAGE: line 81: SyntaxError: 'break' is only valid inside a switch or loop statement
S12.8_A1_T3

PASS Expected parsing failure

TEST COMPLETE
